logs: Add promtail and grafana/loki
Place promtail into the zeek container and configure it to forward
all logs to loki:3100.

Loki consumes in plain text, but could be configured to actually
parse/interpret the logs a bit, but that also works at query time.

Add a new dashboard with dns.log / conn.log. Now need to read the
tutorial for LogQL :-/